scaffolder-templates/wheres-teddy/skeleton/requirements.txt
Flask[async]
gunicorn
requests
pillow
boto3
botocore
